Renamed projects and namespaces to reflect the current name of the project
