--- src/PostgrestSharp.Abstract/Enums/RelationsipType.cs
@@ -1,8 +1,0 @@
-﻿namespace PostgrestSharp.Abstract.Enums;
-public enum RelationsipType
-{
-    OneToOne,
-    OneToMany,
-    ManyToMany,
-    SelfReferencing
-}
