--- src/PostgrestSharp.Abstract/Exceptions/AlreadyExistsException.cs
@@ -1,11 +1,0 @@
-﻿namespace PostgrestSharp.Abstract.Exceptions;
-
-public class AlreadyExistsException(string message)  : Exception
-{
-    
-    
-    public override string ToString()
-    {
-        return message;
-    }
-}
