--- src/PostgrestSharp.Abstract/Interfaces/Services/IJsonObjectHandlerService.cs
@@ -1,9 +1,0 @@
-﻿using System.Text.Json;
-using PostgrestSharp.Abstract.DTOs;
-
-namespace PostgrestSharp.Abstract.Interfaces.Services;
-
-public interface IJsonObjectHandlerService
-{
-    Guid PostMethodHandler(string tableName, JsonElement element);
-}
