--- src/PostgrestSharp.Abstract/Interfaces/Services/PostServices.cs
@@ -1,6 +1,0 @@
-﻿namespace PostgrestSharp.Abstract.Interfaces.Services;
-
-public interface IPostService
-{
-    
-}
