--- src/PostgrestSharp.SQL/postgres_file_ajustado.sql
@@ -1,1 +1,0 @@
-
