--- src/PostgrestSharp.WebApi/PostgrestSharp.WebApi.csproj
@@ -1,24 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net9.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.1"/>
-        <PackageReference Include="Npgsql" Version="9.0.2" />
-        <PackageReference Include="Ulid" Version="1.3.4" />
-        <PackageReference Include="RepoDb.PostgreSql" Version="1.13.1" />
-        <PackageReference Include="SqlKata" Version="2.4.0" />
-        <PackageReference Include="SqlKata.Execution" Version="2.4.0" />
-        
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\PostgrestSharp.Abstract\PostgrestSharp.Abstract.csproj" />
-      <ProjectReference Include="..\PostgrestSharp.Business\PostgrestSharp.Business.csproj" />
-    </ItemGroup>
-
-</Project>
