--- src/PostgrestSharp.WebApi/appsettings.json
@@ -1,9 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Trace",
-      "Microsoft.AspNetCore": "Trace"
-    }
-  },
-  "AllowedHosts": "*"
-}
