+++ src/restsharpdb.Abstract/Enums/RelationsipType.cs
@@ -1,0 +1,8 @@
+﻿namespace PostgrestSharp.Abstract.Enums;
+public enum RelationsipType
+{
+    OneToOne,
+    OneToMany,
+    ManyToMany,
+    SelfReferencing
+}
