+++ src/restsharpdb.Abstract/Exceptions/AlreadyExistsException.cs
@@ -1,0 +1,11 @@
+﻿namespace PostgrestSharp.Abstract.Exceptions;
+
+public class AlreadyExistsException(string message)  : Exception
+{
+    
+    
+    public override string ToString()
+    {
+        return message;
+    }
+}
