+++ src/restsharpdb.Abstract/Interfaces/Services/IJsonObjectHandlerService.cs
@@ -1,0 +1,9 @@
+﻿using System.Text.Json;
+using PostgrestSharp.Abstract.DTOs;
+
+namespace PostgrestSharp.Abstract.Interfaces.Services;
+
+public interface IJsonObjectHandlerService
+{
+    Guid PostMethodHandler(string tableName, JsonElement element);
+}
