+++ src/restsharpdb.Abstract/Interfaces/Services/PostServices.cs
@@ -1,0 +1,6 @@
+﻿namespace PostgrestSharp.Abstract.Interfaces.Services;
+
+public interface IPostService
+{
+    
+}
