+++ src/restsharpdb.Abstract/restsharpdb.Abstract.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net9.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <RootNamespace>PostgrestSharp.Abstract</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Npgsql" Version="9.0.2" />
+      <PackageReference Include="Ulid" Version="1.3.4" />
+    </ItemGroup>
+
+</Project>
