+++ src/restsharpdb.SQL/restsharpdb.SQL.csproj
@@ -1,0 +1,10 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net9.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <RootNamespace>PostgrestSharp.SQL</RootNamespace>
+    </PropertyGroup>
+
+</Project>
