+++ src/restsharpdb.WebApi/appsettings.Development.json
@@ -1,0 +1,12 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Debug",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+
+  "ConnectionStrings": {
+    "PortalDbContext": "server=storage.local;database=<database_name>;persist security info=True;user id=postgres;password=<your password>;Maximum Pool Size=350;Write Buffer Size=16384;Timeout=40;Timezone=-03;Include Error Detail=true"
+  }
+}
