+++ src/restsharpdb.WebApi/appsettings.json
@@ -1,0 +1,9 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Trace",
+      "Microsoft.AspNetCore": "Trace"
+    }
+  },
+  "AllowedHosts": "*"
+}
